Added a FloatBitmap for Convolutional Neural Networks.
I also started to work on a layer setup and a network managment class.

+++ CriaAI-Core/src/util/Random.cpp
@@ -11,7 +11,7 @@ namespace cria_ai
 	}
 	void RandNewSeed()
 	{
-		RandSetSeed(time(0));
+		RandSetSeed((long)time(0));
 	}
 
 	float RandFloat()
